% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Connects to Google Cloud API for user-authentication}
\usage{
authenticate(googleAuthR.client_id, googleAuthR.client_secret)
}
\arguments{
\item{googleAuthR.client_id}{Client id provided by Google API cloud service}

\item{googleAuthR.client_secret}{Client secret provided by Google API cloud service}
}
\value{
a vector of objects
}
\description{
This function can be used to authenticate users
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com}, Yasmine Afify \email{yasmine.afify@cis.asu.edu.eg}, and Nagwa Badr \email{Nagwabadr@cis.asu.edu.eg}
}
