% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLogic.R
\docType{data}
\name{ClassLogic}
\alias{ClassLogic}
\title{Indoor and Outdoor class weights}
\format{
A data frame with 20 rows and 3 variables
}
\usage{
data(ClassLogic)
}
\description{
A dataset containing the objects and its In-door and Out-door weights.
  The variables are as follows:
}
\details{
\itemize{
  \item Object. object name
  \item Class. class of data (Indoor, Outdoor)
  \item Weight. Classification weight (0--1)
}
}
\keyword{datasets}
