% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLabels.R
\name{GetLabels}
\alias{GetLabels}
\title{Connects to Google Cloud API label detection and perform multiple object detection}
\usage{
GetLabels(ImagePaths)
}
\arguments{
\item{ImagePaths}{vector of complete image's paths}
}
\value{
a vector of objects
}
\description{
This function can be used to classify image based on objects detected
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com}, Yasmine Afify \email{yasmine.afify@cis.asu.edu.eg}, and Nagwa Badr \email{Nagwabadr@cis.asu.edu.eg}
}
